+++ app/controllers/application_controller.rb
@@ -2,14 +2,20 @@
 
 # Main web application controller for salsa website
 class ApplicationController < ActionController::Base
-  before_action :configure_permitted_parameters, if: :devise_controller?
-  # Prevent CSRF attacks by raising an exception.
-  # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
+  skip_before_action :verify_authenticity_token, if: :devise_login?
+  before_action :configure_permitted_parameters, if: :devise_controller?
 
   protected
 
+  def devise_login?
+    params[:controller] == 'devise/sessions' && params[:action] == 'create'
+  end
+
   def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:sign_up, keys: [:first_name, :last_name])
+    devise_parameter_sanitizer.permit(
+      :sign_up,
+      keys: [:first_name, :last_name, :email, :password, :password_confirmation]
+    )
   end
 end