+++ app/helpers/application_helper.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-# Default helper for methods used in views
+# Default helper for methods used in views.
 module ApplicationHelper
 end