+++ app/helpers/moves_helper.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-# Helper for models used by moves in views
-module MovesHelper
-end