+++ app/jobs/application_job.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-# Handles long running jobs
+# Handles long running jobs.
 class ApplicationJob < ActiveJob::Base
 end