+++ app/mailers/application_mailer.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Default mailer for application
+# Default mailer for application.
 class ApplicationMailer < ActionMailer::Base
   default from: "#{ENV['website_name']} <#{ActionMailer::Base.smtp_settings[:email]}>"
   layout 'mailer'