+++ config/routes.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
+  root 'external#landing'
+
   scope module: 'external' do
     get :landing
     get :version
@@ -18,12 +20,4 @@ Rails.application.routes.draw do
                                    sign_in: 'login',
                                    sign_out: 'logout' },
                      path: ''
-
-  root to: 'external#landing'
-
-  # For details on the DSL available within this file,
-  # see http://guides.rubyonrails.org/routing.html
-
-  # Serve websocket cable requests in-process
-  # mount ActionCable.server => '/cable'
 end